pyproject.toml
[project]
name = "listen"
version = "0.1.0"
description = "Claude Code skill that turns audio recordings into speaker-labeled transcripts and project-aware meeting notes."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.0.0",
    "pyannote.audio>=3.1",
    "torch>=2.1",
    "torchaudio>=2.1",
    "soundfile>=0.12",
    "numpy>=1.24",
]

[tool.uv]
package = false
# For CUDA torch, install a matching wheel from https://download.pytorch.org/whl/
# after `uv sync` (the default torch on PyPI is the CUDA build on Linux; CPU-only
# users can override with `uv pip install torch --index-url .../whl/cpu`).
